Status to date:
Bots:

Training access permissions:
Roger <ER>, Demonic, WPE <ER>, Randedl, ZeUs!!, Elmo!, WingZero, and all SMODS and TWL-Ops

Done

Time Race:
We need this implemented asap. We also need to know if this will affect database and stat functions as well as connectivity. Will need to make sure they report the flag time to the scoreboard as well.

Working on it

MVP System:
Unless you guys can come up with a mighty convincing and solid system, we are going to stick with what we did last season. We can make adjustments for future TWL's.

Working on it

Bot Announcements:
Here are some bot announcements we should edit or add.

Arena - ship switches, substitutions, lagouts but only when a player uses all lagouts up, additional messages announcing once a team has a full lineup in,others?
Give me a complete list without ambiguity and this will get done in 30 mins, but if none is given I'll just do what is stated so far

Zone - Zoner needs to state Season 7 instead of Season 6
Done

Private - I'm not sure what all of these are
Cant do this if no one knows

Commands:

I'm lost on the score command, I thought it was fixed. Either way, lets remove it entirely or set it as a greet msg, or incorporate the scores checking system from TWD. All other commands should be fine, as they were fully functional last season.

Done: Score command for anyone not ref or Smod is removed

Referee Locking:
Can we get these bots locked to the assigned referee? For instance, once they load the TWL module, the bot is theirs alone, with no one else besides a TWL-Op or SMOD being able to do anything on that bot unless the access is granted to another.

Done

Speccing and Lag:
We need to either decide to use the slow packetloss system, scrap it, or retool it. There are already rules in place for speccing for safety. Lets implement these rules into a bot system so a host does not have to police this.

Get 2d working on this + alias checking

Warp:
Can we add a second !setupwarp for the 30 second wait? Just to be assured there is not hanky panky.
Anything else?

Can someone else do this please