Reply to changeset:1580 ; I put a special character in there so that the bot would send a PM but it wouldn't contain anything since Continuum can't show the special character (ascii 149).
The bot would send an empty PM effectively bypassing String.trim() that keeps you from sending a normal empty PM.

Obviously this gives compilation problems in other IDE's. I re-added the line with a normal character (!`) to keep the spacing in the response.

Maybe a method should be made to send an empty PM or remove the String.trim() from BotAction.sendPrivateMessage().